PyTorch/1.13/attention_is_all_you_need/requirements.txt
torchtext
dill==0.3.4
tqdm
iopath
numpy
spacy==2.3.5
